separate file for status map definition and helpers
