+++ frontend/src/store/data_model/status_map.ts
@@ -1,0 +1,33 @@
+/* Copyright 2020 Google LLC
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    https://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License. */
+
+// We don't want to display API status names directly,
+//  but we still use them in the code
+const internalStatusMap: Record<string, string> = {
+  ACTIVE: "Applicable",
+  CLAIMED: "In progress",
+  SUCCEEDED: "Success",
+  FAILED: "Failed",
+  DISMISSED: "Dismissed"
+};
+
+export function throwIfInvalidStatus(statusName: string): void {
+  if (!(statusName in internalStatusMap))
+    throw `invalid status name passed: ${statusName}`;
+}
+
+export function getInternalStatusMapping(statusName: string): string {
+  throwIfInvalidStatus(statusName);
+  return internalStatusMap[statusName];
+}
